[#52] feat : 회화기능 수정, 키워드 조회기능 수정

+++ backend/src/main/java/com/ssafy/newspeak/conversation/service/ReportService.java
@@ -1,5 +1,8 @@
 package com.ssafy.newspeak.conversation.service;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ssafy.newspeak.conversation.dto.report.ReportCompleteResponse;
 import com.ssafy.newspeak.conversation.dto.report.ReportDto;
 import com.ssafy.newspeak.conversation.dto.report.ReportResponse;
 import com.ssafy.newspeak.conversation.entity.Report;
@@ -23,18 +26,20 @@ public class ReportService {
     private final ReportRepository reportRepository;
     private final UserRepo userRepository;
 
-    public ReportDto create(String content) {
-//        User user = userRepository.findById(userId)
-//                .orElseThrow(RuntimeException::new);
-        Report report = Report.of(content);
+    public ReportDto create(String title, ReportCompleteResponse content, Long userId) throws JsonProcessingException {
+        User user = userRepository.findById(userId)
+                .orElseThrow(RuntimeException::new);
+        ObjectMapper objectMapper = new ObjectMapper();
+        String result = objectMapper.writeValueAsString(content);
+        Report report = Report.of(title, result, user);
         reportRepository.save(report);
 
         return ReportDto.from(report);
     }
 
     @Transactional(readOnly = true)
-    public List<ReportDto> getList() {
-        return reportRepository.findAll().stream()
+    public List<ReportDto> getList(Long userId) {
+        return reportRepository.findAll(userId).stream()
                 .map(ReportDto::from)
                 .toList();
     }
@@ -46,11 +51,14 @@ public class ReportService {
                 .orElseThrow(() -> new NoSuchReportException(reportId));
     }
 
-//    public ReportResponse findByUser
-
-    public void deleteOne(Long reportId) throws NoSuchReportException {
+    public void deleteOne(Long reportId, Long userId) throws NoSuchReportException {
         Report report = reportRepository.findById(reportId)
-                .orElseThrow(RuntimeException::new);
+                .orElseThrow(() -> new NoSuchReportException(reportId));
+
+        if (!report.getUser().getId().equals(userId)) {
+            throw new RuntimeException("You do not have permission to delete this report.");
+        }
+
         reportRepository.delete(report);
     }
 }
